admin user

+++ src/main/java/kg/airbnb/airbnb/services/impl/UserServiceImpl.java
@@ -1,11 +1,15 @@
 package kg.airbnb.airbnb.services.impl;
 
 import kg.airbnb.airbnb.dto.responses.SimpleResponse;
+import kg.airbnb.airbnb.dto.responses.UserBookingsResponse;
 import kg.airbnb.airbnb.dto.responses.UserProfileResponse;
 import kg.airbnb.airbnb.dto.responses.UserResponse;
+import kg.airbnb.airbnb.enums.Role;
 import kg.airbnb.airbnb.exceptions.ForbiddenException;
 import kg.airbnb.airbnb.mappers.UserProfileViewMapper;
+import kg.airbnb.airbnb.mappers.booking.BookingViewMapper;
 import kg.airbnb.airbnb.models.auth.User;
+import kg.airbnb.airbnb.repositories.BookingRepository;
 import kg.airbnb.airbnb.repositories.UserRepository;
 import kg.airbnb.airbnb.services.UserService;
 import org.springframework.security.core.Authentication;
@@ -20,10 +24,10 @@ public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
     private final UserProfileViewMapper viewMapper;
 
-    public UserServiceImpl(UserRepository userRepository, UserProfileViewMapper viewMapper) {
+    public UserServiceImpl(UserRepository userRepository
+            , UserProfileViewMapper viewMapper) {
         this.userRepository = userRepository;
         this.viewMapper = viewMapper;
-
     }
 
     @Override
@@ -40,13 +44,34 @@ public class UserServiceImpl implements UserService {
                 new ForbiddenException("An unregistered user cannot post an ad !"));
     }
 
-    public SimpleResponse delete(Long id) {
-        User user = userRepository.findById(id).get();
-        userRepository.delete(user);
-        return new SimpleResponse("Пользователь успешно удалён!");
+    public SimpleResponse deleteUser(Long id) {
+        User users = getAuthenticatedUser();
+        if (users.getRole().equals(Role.ADMIN)) {
+            User user = userRepository.findById(id).get();
+            userRepository.delete(user);
+            return new SimpleResponse("Пользователь успешно удалён!");
+        } else {
+            throw new ForbiddenException("Only admin can access this page!");
+        }
     }
 
-    public List<UserResponse> getAll() {
-        return UserProfileViewMapper.viewFindAllUser(userRepository.findAll());
+    public List<UserResponse> getAllUser() {
+        User user = getAuthenticatedUser();
+        if (user.getRole().equals(Role.ADMIN)) {
+            return UserProfileViewMapper.viewFindAllUser(userRepository.findAll());
+        } else {
+            throw new ForbiddenException("Only admin can access this page!");
+        }
+    }
+
+    public List<UserBookingsResponse> getAllBookings(Long id) {
+        User user = getAuthenticatedUser();
+        if (user.getRole().equals(Role.ADMIN)) {
+            User users = userRepository.findById(id).get();
+            return BookingViewMapper.viewAllBooking(users.getBookings());
+
+        } else {
+            throw new ForbiddenException("Only admin can access this page!");
+        }
     }
 }
